refactor: inline determine branch
