+++ internal/actions/create/create.go
@@ -180,3 +180,23 @@ func Action(ctx *runtime.Context, opts Options) error {
 
 	return nil
 }
+
+func determineBranch(ctx *runtime.Context, opts *Options, commitMessage string, scope string) (engine.Branch, error) {
+	branchName := opts.BranchName
+	if branchName == "" {
+		// Get pattern from options (always valid, default applied in GetBranchPattern)
+		pattern := opts.BranchPattern
+
+		// Generate branch name from pattern
+		var err error
+		branchName, err = pattern.GetBranchName(ctx.Context, commitMessage, scope)
+		if err != nil {
+			return engine.Branch{}, err
+		}
+	} else {
+		// Sanitize provided branch name
+		branchName = utils.SanitizeBranchName(branchName)
+	}
+
+	return ctx.Engine.GetBranch(branchName), nil
+}
